fix(system_info): use standard `u64` type

+++ src/model/info/system_info/disk.rs
@@ -5,13 +5,13 @@ pub struct DiskInfo {
     #[serde(rename = "mount")]
     mount_point: String,
     #[serde(rename = "total")]
-    total_mb: u128,
+    total_mb: u64,
     #[serde(rename = "free")]
-    free_mb: u128,
+    free_mb: u64,
 }
 
 impl DiskInfo {
-    pub fn new(mount_point: String, total_mb: u128, free_mb: u128) -> Self {
+    pub fn new(mount_point: String, total_mb: u64, free_mb: u64) -> Self {
         Self {
             mount_point,
             total_mb,
